+++ app/src/main/java/com/sola/anime/ai/generator/feature/main/MainActivity.kt
@@ -11,6 +11,7 @@ import com.basic.common.extension.getDimens
 import com.basic.common.extension.lightStatusBar
 import com.basic.common.extension.transparent
 import com.jakewharton.rxbinding2.view.clicks
+import com.revenuecat.purchases.BuildConfig
 import com.revenuecat.purchases.Purchases
 import com.revenuecat.purchases.getCustomerInfoWith
 import com.sola.anime.ai.generator.common.App
@@ -72,19 +73,7 @@ class MainActivity : LsActivity<ActivityMainBinding>(ActivityMainBinding::inflat
     @SuppressLint("SimpleDateFormat")
     private fun syncUserPurchased() {
         Purchases.sharedInstance.getCustomerInfoWith { customerInfo ->
-            customerInfo.entitlements.all.forEach {
-                Timber.tag("Main12345").e("Key: ${it.key} --- ${it.value.isActive}")
-            }
-            customerInfo.latestExpirationDate?.let { date ->
-                Timber.tag("Main12345").e("Time Expired: ${SimpleDateFormat("dd/MM/yyyy - hh:mm:ss").format(date)}")
-            }
-            for (i in customerInfo.allPurchaseDatesByProduct){
-                i.value?.let { date ->
-                    Timber.tag("Main12345").e("Product: ${i.key} --- Purchased: ${SimpleDateFormat("dd/MM/yyyy - hh:mm:ss").format(date)}")
-                }
-            }
             val isActive = customerInfo.entitlements["premium"]?.isActive ?: false
-            Timber.tag("Main12345").e("Premium is active: $isActive")
 
             if (isActive){
                 customerInfo
@@ -92,48 +81,50 @@ class MainActivity : LsActivity<ActivityMainBinding>(ActivityMainBinding::inflat
                     .filter { it.value != null }
                     .filter { it.key.contains(Constraint.Iap.SKU_WEEK) || it.key.contains(Constraint.Iap.SKU_MONTH) ||it.key.contains(Constraint.Iap.SKU_YEAR) }
                     .takeIf { it.isNotEmpty() }
-                    ?.maxBy { it.value!! }
+                    ?.maxByOrNull { it.value!! }
                     ?.let { map ->
-                        if (map.value == null) return@let
-
                         val latestPurchasedProduct = map.key
-                        val timeExpired = when {
-                            prefs.isUpgraded.get() -> when {
-                                latestPurchasedProduct.contains(Constraint.Iap.SKU_WEEK) -> 604800016L // 1 Week
-                                latestPurchasedProduct.contains(Constraint.Iap.SKU_MONTH) -> 2629800000L // 1 Month
-                                latestPurchasedProduct.contains(Constraint.Iap.SKU_YEAR) -> 31557600000L // 1 Year
-                                else -> 21600000L // 6 Hour
-                            }
-                            else -> 21600000L // 6 Hour
+                        val latestDatePurchased = map.value ?: return@let
+                        val expiredDate = customerInfo.getExpirationDateForProductId(latestPurchasedProduct) ?: return@let
+
+                        val expiredDateTime = when {
+                            prefs.isUpgraded.get() -> expiredDate.time
+                            else -> latestDatePurchased.time + 21600000L // Day time purchased + 6 hours
                         }
 
-                        val differenceInMillis = timeExpired - (Date().time - map.value!!.time)
-                        val days = TimeUnit.MILLISECONDS.toDays(differenceInMillis)
-                        val hours = TimeUnit.MILLISECONDS.toHours(differenceInMillis) % 24
-                        val minutes = TimeUnit.MILLISECONDS.toMinutes(differenceInMillis) % 60
-                        val seconds = TimeUnit.MILLISECONDS.toSeconds(differenceInMillis) % 60
+                        val differenceInMillis = expiredDateTime - Date().time
+                        if (differenceInMillis > 0){
+                            val days = TimeUnit.MILLISECONDS.toDays(differenceInMillis)
+                            val hours = TimeUnit.MILLISECONDS.toHours(differenceInMillis) % 24
+                            val minutes = TimeUnit.MILLISECONDS.toMinutes(differenceInMillis) % 60
+                            val seconds = TimeUnit.MILLISECONDS.toSeconds(differenceInMillis) % 60
 
-                        when {
-                            days >= 0 && hours >= 0 && minutes >= 0 && seconds > 0 -> {
-                                prefs.isUpgraded.set(true)
-                                prefs.timeExpiredIap.set(differenceInMillis)
+                            when {
+                                days >= 0 && hours >= 0 && minutes >= 0 && seconds > 0 -> {
+                                    prefs.isUpgraded.set(true)
+                                    prefs.timeExpiredIap.set(differenceInMillis)
+                                }
+                                else -> {
+                                    prefs.isUpgraded.delete()
+                                    prefs.timeExpiredIap.delete()
+                                }
                             }
-                            else -> {
-                                prefs.isUpgraded.set(false)
-                                prefs.timeExpiredIap.delete()
-                            }
+
+                            Timber.tag("Main12345").e("##### MAIN #####")
+                            Timber.tag("Main12345").e("Time Purchased: ${SimpleDateFormat("dd/MM/yyyy - hh:mm:ss").format(latestDatePurchased)}")
+                            Timber.tag("Main12345").e("Time Expired: ${SimpleDateFormat("dd/MM/yyyy - hh:mm:ss").format(expiredDate)}")
+                            Timber.tag("Main12345").e("Date: $days --- $hours:$minutes:$seconds")
                         }
-
-                        Timber.tag("Main12345").e("Time Purchased: ${SimpleDateFormat("dd/MM/yyyy - hh:mm:ss").format(map.value!!)}")
-                        Timber.tag("Main12345").e("Date Expired: $days --- $hours:$minutes:$seconds")
                     }
-
                 when {
                     prefs.isUpgraded.get() && !DateUtils.isToday(prefs.latestTimeCreatedArtwork.get()) -> {
                         prefs.numberCreatedArtwork.delete()
                         prefs.latestTimeCreatedArtwork.delete()
                     }
                 }
+            } else {
+                prefs.isUpgraded.delete()
+                prefs.timeExpiredIap.delete()
             }
         }
     }
@@ -174,11 +165,25 @@ class MainActivity : LsActivity<ActivityMainBinding>(ActivityMainBinding::inflat
                 subjectTabClicks.onNext(0)
             }
 
+        prefs
+            .isUpgraded
+            .asObservable()
+            .filter { isUpgraded -> isUpgraded && !prefs.isShowedWaringPremiumDialog.get() }
+            .delay(1, TimeUnit.SECONDS)
+            .observeOn(AndroidSchedulers.mainThread())
+            .subscribeOn(AndroidSchedulers.mainThread())
+            .autoDispose(scope())
+            .subscribe { isUpgraded ->
+                when {
+                    isUpgraded && !prefs.isShowedWaringPremiumDialog.get() -> warningPremiumDialog.show(this)
+                }
+            }
+
         Observable
             .interval(1, TimeUnit.SECONDS)
             .filter { prefs.isUpgraded.get() }
             .map { prefs.timeExpiredIap.get() }
-            .filter { differenceInMillis -> differenceInMillis != -1L }
+            .filter { differenceInMillis -> differenceInMillis != -1L && differenceInMillis != -2L }
             .observeOn(AndroidSchedulers.mainThread())
             .subscribeOn(AndroidSchedulers.mainThread())
             .autoDispose(scope())
@@ -192,10 +197,13 @@ class MainActivity : LsActivity<ActivityMainBinding>(ActivityMainBinding::inflat
                     days >= 0 && hours >= 0 && minutes >= 0 && seconds > 0 -> {
                         prefs.timeExpiredIap.set(differenceInMillis - 1000L)
                     }
-                    else ->  prefs.isUpgraded.set(false)
+                    else ->  {
+                        prefs.isUpgraded.delete()
+                        prefs.timeExpiredIap.delete()
+                    }
                 }
 
-                Timber.tag("Main12345").e("Date Expired: $days --- $hours:$minutes:$seconds")
+                Timber.tag("Main12345").e("Date: $days --- $hours:$minutes:$seconds")
             }
     }
 
@@ -219,10 +227,6 @@ class MainActivity : LsActivity<ActivityMainBinding>(ActivityMainBinding::inflat
                 this.addFragment(fragments = fragments.toTypedArray())
             }
             this.offscreenPageLimit = this.adapter?.count ?: 0
-        }
-
-        when {
-            prefs.isUpgraded.get() && !prefs.isShowedWaringPremiumDialog.get() -> warningPremiumDialog.show(this)
         }
     }
